+++ package.json
@@ -8,6 +8,7 @@
     "@emotion/styled": "^11.11.0",
     "@mui/material": "^5.13.5",
     "@reduxjs/toolkit": "^1.9.5",
+    "@smakss/react-scroll-direction": "^4.0.3",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
