+++ src/Components/EmailForm/EmailForm.css
@@ -21,7 +21,7 @@
 
 }
 
-.contact-form {
+.contactForm {
     padding: 5px;
     /* max-width: 800px; */
 
@@ -37,7 +37,7 @@
 }
 
 
-.form-group {
+.formGroup {
     margin-bottom: 20px;
     color: #fff;
