--- src/Components/Footer/FooterComponent.css
@@ -1,75 +1,0 @@
-/* .footer {
-    background-color: rgba(67, 0, 0, 0.698);
-  
-    padding: 30px 0;
-}
-
-.footer__content {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.footer__logo img {
-    max-height: 50px;
-}
-
-.footer__links ul {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-    display: flex;
-}
-
-.footer__list-item {
-    margin-right: 15px;
-}
-
-.footer__list-item:last-child {
-    margin-right: 0;
-}
-
-.footer__list-item a {
-    color: #555555;
-    text-decoration: none;
-}
-
-.footer__list-item a:hover {
-    text-decoration: underline;
-}
-
-.footer__social-media a {
-    color: #555555;
-    margin-right: 10px;
-    font-size: 20px;
-}
-
-.footer__social-media a:hover {
-    color: #333333;
-}
-
-.footer__bottom {
-    height: 60px;
-    background-color: #2c3531;
-    padding: 15px 0;
-    margin-top: 30px;
-}
-
-.footer__bottom p {
-    margin: 0;
-    color: #777777;
-    text-align: center;
-    font-size: 14px;
-}
-
-.footerIcons {
-    width: 30px;
-    margin-top: 10px;
-    transition: 1s;
-}
-
-.footerIcons:hover {
-    width: 45px;
-    z-index: 6;
-    transition: .5s;
-} */
