+++ src/Components/MainProfolioPage/MainPageComponent.css
@@ -1,3 +1,11 @@
-.MainPageWrapper {
-    margin-top: 100px;
+.MainPageWrapper{
+    width:100vw;
+    height:100vh;
+ display: flex;
+ flex-direction: row;
+}
+
+.sideBarPositionWrapper{
+    width:30vw;
+    
 }
