--- src/Components/NavBar/NavBarComponent.css
@@ -1,139 +1,0 @@
-.navbar {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-
-    /* background-color: #f2f2f2; */
-    /* background-color: #0700006e; */
-    background-color: rgba(67, 0, 0, 0.698);
-    /* background-color: rgba(107, 22, 22, 0.698); */
-    padding: 10px;
-    transition: transform 0.3s;
-    z-index: 100;
-    /* border-radius: 15%; */
-
-}
-
-.testDesign {
-    margin-top: -15px;
-    margin-left: -40px;
-    margin-bottom: -10px;
-    border: 6px solid #111112ab;
-    display: flex;
-    justify-content: left;
-    background-color: #111112ab;
-    border-radius: 60px;
-
-
-    width: 30vmax;
-
-
-}
-
-.logoDesign {
-    position: relative;
-    top: -1px;
-    margin-bottom: -10px;
-    margin-top: -2px;
-    margin-left: 50px;
-    border: 5px solid rgba(0, 0, 0, 0.582);
-    border-radius: 100px;
-    width: 70px;
-    height: 70px;
-}
-
-.logoDesign:hover {
-    background-color: rgba(0, 0, 0, 0.826);
-}
-
-.navbar.visible {
-    transform: translateY(0);
-}
-
-.navbar.hidden {
-    transform: translateY(-100%);
-}
-
-
-/* .navbar {
-    background-color: #f2f2f2;
-    padding: 10px;
-}
-*/
-.navbar-container {
-    display: flex;
-    justify-content: space-between;
-    margin-right: 100px !important;
-    align-items: center;
-    /* max-width: 960px; */
-    margin: 0 auto;
-}
-
-.navbar-logo {
-    color: #ffffff;
-    font-weight: bold;
-    text-decoration: none;
-    font-size: 20px;
-}
-
-.navbar-menu {
-    list-style-type: none;
-    display: flex;
-    margin-right: 30px;
-}
-
-.navbar-item {
-    margin-left: 20px;
-
-}
-
-.navbar-link {
-    color: #ffffff;
-
-    text-decoration: none;
-    font-size: 16px;
-}
-
-.navbar-link:hover {
-    color: #ffcb9a !important;
-}
-
-.navbar-link:hover {
-    color: #007bff;
-}
-
-@media only screen and (max-width: 400px) {
-    .navbar-container {
-        display: flex;
-        justify-content: left;
-        /* margin-left: 30px; */
-        height: 40px;
-    }
-
-    .navbar-link {
-        font-size: 10px;
-
-
-    }
-
-    .navbar-menu {
-        position: relative;
-        right: 40px;
-
-    }
-
-    .testDesign {
-        width: 50vmax;
-        margin-right: 20px;
-    }
-
-    .logoDesign {
-        margin-top: -4px;
-        margin-left: 40px;
-        width: 50px;
-        height: 50px;
-
-    }
-
-}
