+++ src/Components/Profile/ProfileComponent.module.css
@@ -2,33 +2,33 @@
 html{
     margin:-100 auto;
 } */
-
+.gridDesign.MuiPaper-root{
+    height:100vh;
+    
+}
 .portfolioHolder{
    background-color: rgba(240, 248, 255, 0);
-    position: absolute;
-    top:0px;
-    /* border:2px solid white; */
-    /* margin-bottom:1px auto; */
-    
-
-    width:30vw;
+   border-top-right-radius: 180px;
+   border-bottom-left-radius: 180px;
+   width:400px;
     height:100vh;
     /* display: flex;
     justify-content: center; */
 }
 .ProfileImageHolder{
-    margin-top: 100px;
+    margin-top: 30px;
 
     /* border:5px solid white; */
+    max-width: 100%;
     border-radius: 50%;
-    width:50%;
-    height:30%;
+    width:200px;
+    /* height:30%; */
 }
 .nameTextStyle{
     color:lightgray;
 }
 .navigationSelection{
-    margin-left:5vw;
+    margin-left:100px;
     text-align:left;
 
 }
@@ -38,23 +38,19 @@ html{
     color: lightgray;
 }
 .ScrollItemsContainer{
-    position: absolute;
-    top:0px;
-    right:0px;
-    /* display: flex;
-    float: right; */
-    /* border:3px solid white; */
-    width:69vw;
-    height:100vh;
+    display: flex;
+    flex:1;
+    justify-content: center;
+   
 
 }
-.footer__social-media a {
+.footerSocialMedia a {
     color: #555555;
     margin-right: 10px;
     font-size: 20px;
 }
 
-.footer__social-media a:hover {
+.footerSocialMedia a:hover {
     color: #333333;
 }
 
@@ -73,9 +69,9 @@ html{
 
 @media only screen and (max-width: 1000px) {
     .ProfileImageHolder{
-       
+/*        
         width:50%;
-        height:15%;
+        height:15%; */
     }
     .selectionItem{
         font-size:20px;
