--- src/Components/Profile/ScrollItem.css
@@ -1,8 +1,0 @@
-.aboutMeItem{
-    width: 50vw;
-   text-align: left;
-   display: flex;
-   float:right;
-   color:lightgray;
-   margin:30px;
-}
