+++ src/Components/Profile/ScrollItems.js
@@ -1,47 +1,38 @@
-import React from 'react'
-import './ScrollItem.css'
+import React, {useEffect, useState} from 'react'
+import styles from './ScrollItem.module.css'
+import AboutMe from '../../Content/AboutMe';
+import useDetectScroll from '@smakss/react-scroll-direction';
 
 const ScrollItems = () => {
+    const { scrollDir, scrollPosition } = useDetectScroll()
+    const [scrollUp, setScrollUp] = useState(false);
+    useEffect(() => {
+      const scrollAction = (e) => {
+        // console.log('scroll has been attempted')
+        // console.log('scrollDir: ', scrollDir)
+        // console.log( e.deltaX)
+        if(!scrollUp && e.deltaX<0){
+            setScrollUp(true);
+            console.log('scrolled up')
+        }else if( e.deltaX === 0){
+          setScrollUp(false)
+        }
+      }
+      window.addEventListener('wheel', (e) => scrollAction(e))
+      return () => {
+        window.removeEventListener('wheel', (e) => scrollAction(e))
+      }
+    })
+  
+  
+  
+  
+  
+  
     return (
-        <div>
-            <h1 className="aboutMeItem">About Me:</h1>
-            <h5 className="aboutMeItem">
-                Hey there! I'm Dennis Enwiya, and I'm thrilled to share my
-                passion for software development with you. I have always been
-                captivated by the world of technology, and pursuing a career in
-                this field has been a dream come true. To kickstart my journey,
-                I enrolled in Galvanize, a prestigious 6-month boot camp
-                dedicated to software development. This intense program provided
-                me with invaluable certifications, helping me establish a strong
-                foundation in the field. During my time at Galvanize, I not only
-                sharpened my technical skills but also learned the importance of
-                collaboration and innovative problem-solving.
-                <br />
-                <br />
-                Fueling my desire for knowledge, I decided to pursue a degree in
-                Computer Science from Southern New Hampshire University. This
-                academic experience expanded my understanding of core computer
-                science principles, algorithms, and software development
-                methodologies. It also instilled in me a lifelong passion for
-                learning and staying up to date with the latest technological
-                advancements. Equipped with certifications from Galvanize and a
-                degree from Southern New Hampshire University, I am excited to
-                make a meaningful impact in the software development industry. I
-                thrive on tackling complex coding challenges, building robust
-                applications, and contributing to the continuous innovation that
-                drives the tech landscape forward.
-                <br />
-                <br />
-                I'm an avid believer in personal and professional growth,
-                constantly seeking new opportunities to expand my skill set.
-                With my determination, drive, and solid foundation in software
-                development, I am confident in my ability to bring value to any
-                organization looking for a dedicated and accomplished software
-                developer. Join me on this exciting journey as I strive to make
-                a lasting impact in the ever-evolving world of software
-                development. Together, let's push boundaries and create
-                innovative solutions that shape the future.
-            </h5>
+        <div className={styles.scrollItemsWrapper}>
+            <AboutMe/>
+            
         </div>
     )
 }
